personal-model/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2

# NATS JetStream
nats-py==2.9.0

# Redis async
redis==5.2.0

# PostgreSQL async + pgvector
asyncpg==0.30.0

# HTTP client for LLM calls
httpx==0.27.2

# Utilities
python-dotenv==1.0.1
